WIP: Craete new NTP climate impact section
